cos/queue_ops: extract push + pop ops into siblings (#1034 P3)
Two more sibling submodules under cos/queue_ops/:

**push.rs** (211 LOC) — enqueue + push-front operations:
- cos_queue_push_back (standard enqueue)
- cos_queue_push_front (rollback-of-popped-item path used by worker
  drain when a popped item must be returned to the head before
  draining-side bookkeeping settled)

**pop.rs** (151 LOC) — dequeue path including pop-snapshot stack:
- cos_queue_pop_front (with snapshot for batched rollback)
- cos_queue_pop_front_no_snapshot (non-batched variant)
- cos_queue_pop_front_inner (shared implementation, private)

queue_ops/mod.rs production LOC (excluding mod tests):
  4,203 → 3,865 (Phase 3 of ~5 — combined push and pop in one PR
  for higher throughput now that the directory module + smaller
  helper extractions in P1/P2 have validated the pattern)

Visibility:
- All four pub fns re-exported at pub(in crate::afxdp) to preserve
  cos/mod.rs:20's external paths
- cos_queue_pop_front_inner kept private (only push.rs / pop.rs
  callers use it)

Phase 4+ target: split the test mod block (which is ~3,500 LOC of
the remaining queue_ops/mod.rs raw LOC) by theme (mqfq_*, flow_fair_*,
exact_*, cos_queue_*) into sibling test files, colocated with their
production code per modularity-discipline rule.

build clean, 865 Rust tests pass. Cluster smoke required (touches
the worker push/pop hot path).
