+++ userspace-dp/src/afxdp/cos/queue_ops/mod.rs
@@ -28,6 +28,12 @@ pub(in crate::afxdp) use drain::{
     cos_queue_clear_orphan_snapshot_after_drop, cos_queue_drain_all, cos_queue_restore_front,
 };
 
+// #1034 P3: push + pop ops split into siblings.
+mod pop;
+mod push;
+pub(in crate::afxdp) use pop::{cos_queue_pop_front, cos_queue_pop_front_no_snapshot};
+pub(in crate::afxdp) use push::{cos_queue_push_back, cos_queue_push_front};
+
 
 
 #[inline]
@@ -91,354 +97,10 @@ pub(in crate::afxdp) fn cos_queue_front(queue: &CoSQueueRuntime) -> Option<&CoSP
     queue.flow_bucket_items[bucket].front()
 }
 
-#[inline]
-pub(in crate::afxdp) fn cos_queue_push_back(queue: &mut CoSQueueRuntime, item: CoSPendingTxItem) {
-    let item_len = cos_item_len(&item);
-    let flow_key = cos_item_flow_key(&item);
-    // #774: maintain local_item_count alongside the queue pushes
-    // so cos_queue_accepts_prepared becomes O(1). `matches!` on a
-    // tagged enum is a single branch; far cheaper than an O(n)
-    // scan at check time.
-    if matches!(item, CoSPendingTxItem::Local(_)) {
-        queue.local_item_count = queue.local_item_count.saturating_add(1);
-    }
-    // #785 Phase 3 — Codex round-3 HIGH + NEW-1: any push_back
-    // invalidates every outstanding pop snapshot. A subsequent
-    // push_front must re-anchor fresh rather than restoring
-    // pre-pop head/tail of a bucket whose state has since changed
-    // underneath us. Cleared in bulk (not per-bucket) because the
-    // cost of a tiny Vec::clear is ~zero and the safety contract is
-    // simpler: after any new enqueue, no rollback can use ANY
-    // snapshot captured before it.
-    queue.pop_snapshot_stack.clear();
-    account_cos_queue_flow_enqueue(queue, flow_key, item_len);
-    if !queue.flow_fair {
-        queue.items.push_back(item);
-        return;
-    }
-    let bucket = cos_flow_bucket_index(queue.flow_hash_seed, flow_key);
-    let bucket_queue = &mut queue.flow_bucket_items[bucket];
-    let was_empty = bucket_queue.is_empty();
-    bucket_queue.push_back(item);
-    if was_empty {
-        queue.flow_rr_buckets.push_back(bucket as u16);
-    }
-}
 
-#[inline]
-pub(in crate::afxdp) fn cos_queue_push_front(queue: &mut CoSQueueRuntime, item: CoSPendingTxItem) {
-    let item_len = cos_item_len(&item);
-    let flow_key = cos_item_flow_key(&item);
-    if matches!(item, CoSPendingTxItem::Local(_)) {
-        queue.local_item_count = queue.local_item_count.saturating_add(1);
-    }
-    if !queue.flow_fair {
-        account_cos_queue_flow_enqueue(queue, flow_key, item_len);
-        queue.items.push_front(item);
-        return;
-    }
-    let bucket = cos_flow_bucket_index(queue.flow_hash_seed, flow_key);
-    // #913: peek-then-pop snapshot consumption.
-    //
-    // Three states:
-    //   1. Empty stack: legitimate (drain_all cleared it; or
-    //      fresh-flow / non-Phase-3 caller). Aggregate-bytes
-    //      rewind path — `vtime -= item_len` pairs with the
-    //      no-snapshot pop's `vtime += bytes` for round-trip
-    //      neutrality (see plan §3.7 walkthrough for
-    //      drain_all→restore_front).
-    //   2. Top entry's bucket matches: hot-path matched
-    //      rollback. Pop and restore vtime + head/tail
-    //      from snapshot (closes #913 max-based advance).
-    //   3. Top entry's bucket DOES NOT match: hard contract
-    //      violation. With §3.4's scratch-builder orphan
-    //      cleanup in place, this is believed unreachable in
-    //      current code. `assert!(false)` panics in BOTH dev
-    //      and release.
-    //
-    //      No supervisor in this PR (#913 R4 revert): the
-    //      panic propagates to the default Rust panic
-    //      handler, which emits the panic message to stderr
-    //      → journald and kills the worker thread. The
-    //      helper process keeps running with one fewer
-    //      worker; bindings served by that worker stall
-    //      until the daemon is restarted via config change
-    //      or operator intervention. SAME blast radius as
-    //      every existing `unwrap`/`expect`/`panic!` site
-    //      in `worker_loop` — #913 introduces zero
-    //      incremental panic risk. Cross-cutting panic
-    //      supervision (catch_unwind on helper side +
-    //      parent-side restart in xpfd) tracked in #925.
-    let stack_top_bucket = queue
-        .pop_snapshot_stack
-        .last()
-        .map(|s| usize::from(s.bucket));
-    let snapshot = match stack_top_bucket {
-        None => None,
-        Some(top) if top == bucket => queue.pop_snapshot_stack.pop(),
-        Some(top) => {
-            assert!(
-                false,
-                "pop_snapshot_stack bucket mismatch on push_front: \
-                 top entry's bucket {} != target bucket {}; a \
-                 caller pop+dropped an item without §3.4 cleanup, \
-                 or violated the pop→push_front-same-item contract",
-                top, bucket,
-            );
-            unreachable!()
-        }
-    };
 
-    // #913: vtime restore — symmetric inverse of the §3.1 advance.
-    // Matched-snapshot path: restore from snapshot for both the
-    // was_empty (drained-bucket) and active-bucket branches.
-    // Empty-stack path: legacy aggregate-bytes rewind paired with
-    // the no-snapshot pop's `vtime += bytes`.
-    match snapshot.as_ref() {
-        Some(snap) => {
-            queue.queue_vtime = snap.pre_pop_queue_vtime;
-        }
-        None => {
-            queue.queue_vtime = queue.queue_vtime.saturating_sub(item_len);
-        }
-    }
-    // #917 Phase 3: republish the rolled-back queue_vtime so peers
-    // see the restored value, not the speculative pop's advanced
-    // value. Without this, a peer reading mid-rollback would see
-    // an inflated V_min slot for this worker — over-throttling
-    // peers until the next pop fixes it.
-    if let Some(floor) = queue.vtime_floor.as_ref() {
-        if let Some(slot) = floor.slots.get(queue.worker_id as usize) {
-            slot.publish(queue.queue_vtime);
-        }
-    }
 
-    let was_empty = queue.flow_bucket_items[bucket].is_empty();
-    if was_empty {
-        // Bucket was drained by the matching pop. Snapshot (if
-        // present) holds the exact pre-pop head/tail so we can
-        // restore them.
-        if let Some(snap) = snapshot {
-            queue.flow_bucket_bytes[bucket] =
-                queue.flow_bucket_bytes[bucket].saturating_add(item_len);
-            queue.flow_bucket_head_finish_bytes[bucket] = snap.pre_pop_head_finish;
-            queue.flow_bucket_tail_finish_bytes[bucket] = snap.pre_pop_tail_finish;
-            queue.active_flow_buckets = queue.active_flow_buckets.saturating_add(1);
-            if queue.active_flow_buckets > queue.active_flow_buckets_peak {
-                queue.active_flow_buckets_peak = queue.active_flow_buckets;
-            }
-            queue.flow_bucket_items[bucket].push_front(item);
-            queue.flow_rr_buckets.push_front(bucket as u16);
-            return;
-        }
-        // No snapshot — drain_all/restore_front path or fresh-flow
-        // caller. Standard idle-bucket re-anchor.
-        // The aggregate-bytes vtime rewind above leaves vtime
-        // correctly positioned for `max(tail, vtime) + bytes`
-        // (see plan §3.7 walkthrough for the drain_all case).
-        account_cos_queue_flow_enqueue(queue, flow_key, item_len);
-        queue.flow_bucket_items[bucket].push_front(item);
-        queue.flow_rr_buckets.push_front(bucket as u16);
-        return;
-    }
-    // #785 Phase 3 — MQFQ push_front onto an ACTIVE bucket.
-    //
-    // Codex adversarial review (round-2) flagged this path as HIGH:
-    // the prior revision funnelled through
-    // `account_cos_queue_flow_enqueue`, which only advances `tail`
-    // on an active bucket — head stayed stale at a value keyed off
-    // whatever was the HEAD packet before this push_front.
-    // Selection would then pick the bucket based on the STALE head
-    // finish (stale because the item-queue front changed), and the
-    // subsequent non-drain pop would `head += bytes(next_head)`
-    // off the stale base, producing arbitrary finish values.
-    //
-    // Fix: push_front is only called from TX-ring-full restoration
-    // paths where an item was JUST popped from this same bucket.
-    // We reverse that pop's head-advance: at pop time we computed
-    // `head += bytes(what_is_now_front)`. At push_front time we
-    // subtract the SAME quantity to get back to the pop-time head
-    // (which was the popped item's finish). The restored item
-    // takes over as the new head and inherits that finish — which
-    // is exactly what it had before the pop. Net effect: the
-    // pop-and-restore round-trip is finish-time neutral, which is
-    // what correctness on the error-retry path demands.
-    //
-    // #913: vtime is already restored above (snapshot path or
-    // aggregate-bytes path). The active-bucket head reversal
-    // here is unchanged from pre-#913 — `head -= bytes(current_head)`
-    // is correct under MQFQ "drops consume virtual service"
-    // semantics. Reasoning:
-    //
-    // - Single-pop case: push_front is the exact inverse of the
-    //   most recent pop. head was advanced by bytes(current_head);
-    //   subtracting reverses it.
-    // - Multi-pop case with mid-Drop (e.g., pop A1, pop A2, drop A2,
-    //   restore A1 while A3 is in bucket): head=4500 after pop A2.
-    //   Arithmetic gives head=4500-bytes(A3=1500)=3000. Subsequent
-    //   pop A1 then advances head to 3000+bytes(A3)=4500. A3 ends
-    //   up at finish=4500, preserving A2's "consumed virtual
-    //   service" — competing buckets between 3000 and 4500
-    //   correctly drain before A3.
-    //
-    // (Codex code-review R8 initially flagged this as wrong with
-    // recommendation to use snap.pre_pop_head_finish; R9 then
-    // reversed when its own walkthrough showed the arithmetic
-    // result is needed for the post-restore-pop case. Documented
-    // in §3.3 of the plan.)
-    let current_head_bytes = queue.flow_bucket_items[bucket]
-        .front()
-        .map(cos_item_len)
-        .unwrap_or(0);
-    queue.flow_bucket_head_finish_bytes[bucket] = queue.flow_bucket_head_finish_bytes[bucket]
-        .saturating_sub(current_head_bytes);
-    queue.flow_bucket_bytes[bucket] = queue.flow_bucket_bytes[bucket].saturating_add(item_len);
-    queue.flow_bucket_items[bucket].push_front(item);
-}
 
-#[inline]
-pub(in crate::afxdp) fn cos_queue_pop_front(queue: &mut CoSQueueRuntime) -> Option<CoSPendingTxItem> {
-    cos_queue_pop_front_inner(queue, true)
-}
-
-/// #785 Phase 3 — Codex round-3 NEW-2 / Rust reviewer LOW:
-/// teardown-only variant of `cos_queue_pop_front` that does NOT
-/// push a rollback snapshot. Used by drain-all-items-until-empty
-/// paths (`cos_queue_drain_all` and the worker teardown loop)
-/// where the drained items are either discarded or restored via
-/// a single reverse push_front loop that doesn't need per-pop
-/// pre-state capture (nothing has mutated the bucket between
-/// drain and restore in those paths).
-///
-/// Without this variant, a teardown of >TX_BATCH_SIZE items would
-/// grow `pop_snapshot_stack` past its documented bound and trip
-/// the per-pop debug_assert.
-#[inline]
-pub(in crate::afxdp) fn cos_queue_pop_front_no_snapshot(
-    queue: &mut CoSQueueRuntime,
-) -> Option<CoSPendingTxItem> {
-    cos_queue_pop_front_inner(queue, false)
-}
-
-#[inline]
-fn cos_queue_pop_front_inner(
-    queue: &mut CoSQueueRuntime,
-    push_snapshot: bool,
-) -> Option<CoSPendingTxItem> {
-    let item = if !queue.flow_fair {
-        queue.items.pop_front()?
-    } else {
-        // #785 Phase 3 — MQFQ: pop from the bucket whose head
-        // packet has the smallest virtual-finish-time, not DRR
-        // rotation order. The active set (`flow_rr_buckets`) is
-        // still maintained on 0↔>0 transitions so the min-scan
-        // only iterates the currently-active buckets (typically
-        // 2-16), not all 1024.
-        let bucket_u16 = cos_queue_min_finish_bucket(queue)?;
-        let bucket = usize::from(bucket_u16);
-        if push_snapshot {
-            // #785 Phase 3 — Codex round-3 HIGH + NEW-1: snapshot
-            // pre-pop bucket + vtime state BEFORE we mutate anything,
-            // and push onto the per-queue LIFO stack. Every popped
-            // item gets its own snapshot so a batched rollback (N
-            // pops into scratch, submit a prefix, push_front the tail
-            // in LIFO order) can restore exact pre-pop head/tail for
-            // EVERY item — not just the most recent pop.
-            //
-            // Earlier revision kept a single `Option<...>`; Codex
-            // NEW-1 flagged that earlier drained buckets in a
-            // multi-pop rollback fell back to the
-            // `max(tail, queue_vtime) + bytes` re-anchor formula,
-            // which can overshoot the pre-pop head when queue_vtime
-            // has advanced since the bucket's original enqueue.
-            //
-            // Stack capacity is preallocated to TX_BATCH_SIZE
-            // (see types.rs), so this push is amortized O(1) and
-            // allocation-free.
-            //
-            // #785 Phase 3 — Codex round-3 NEW-2 / Rust reviewer LOW:
-            // debug_assert the stack stays within its documented bound.
-            // Drain helpers clear at batch start and teardown paths
-            // use `cos_queue_pop_front_no_snapshot`. If this trips
-            // under dev/test, a new caller is leaking snapshots
-            // and could realloc on the hot path in release builds.
-            debug_assert!(
-                queue.pop_snapshot_stack.len() < TX_BATCH_SIZE,
-                "pop_snapshot_stack exceeded TX_BATCH_SIZE bound ({}); \
-                 a caller is leaking snapshots — drain helpers must \
-                 clear at batch start and teardown paths must use \
-                 cos_queue_pop_front_no_snapshot",
-                TX_BATCH_SIZE,
-            );
-            queue.pop_snapshot_stack.push(CoSQueuePopSnapshot {
-                bucket: bucket_u16,
-                pre_pop_head_finish: queue.flow_bucket_head_finish_bytes[bucket],
-                pre_pop_tail_finish: queue.flow_bucket_tail_finish_bytes[bucket],
-                pre_pop_queue_vtime: queue.queue_vtime,
-            });
-        }
-        // #913: capture served_finish (the popped packet's finish
-        // time) BEFORE pop_front + head-advance below mutate it.
-        let served_finish = queue.flow_bucket_head_finish_bytes[bucket];
-        let item = queue.flow_bucket_items[bucket].pop_front()?;
-        // #913: branched vtime advance.
-        // - push_snapshot=true (hot path / `cos_queue_pop_front`):
-        //   MQFQ served-finish semantics — `vtime = max(vtime,
-        //   served_finish)`. Closes #911 same-class HOL by
-        //   tracking the system frontier (smallest head_finish
-        //   across active buckets at pop time) instead of
-        //   aggregate bytes.
-        // - push_snapshot=false (`cos_queue_pop_front_no_snapshot`,
-        //   used by drain_all + worker.rs:1859 teardown):
-        //   legacy `vtime += bytes` retained. The
-        //   `demote_prepared_cos_queue_to_local` failure-restore
-        //   path (drain_all → restore_front) relies on this
-        //   symmetry with push_front's `vtime -= item_len`
-        //   rewind for round-trip neutrality. drain_all clears
-        //   the snapshot stack at start so push_front of the
-        //   restored items takes the empty-stack aggregate
-        //   path. See plan §3.5 / §3.7.
-        if push_snapshot {
-            queue.queue_vtime = queue.queue_vtime.max(served_finish);
-        } else {
-            let bytes = cos_item_len(&item);
-            queue.queue_vtime = queue.queue_vtime.saturating_add(bytes);
-        }
-        // #940: V_min publish moved to post-settle commit boundary.
-        // See `publish_committed_queue_vtime` for details.
-        if let Some(next_head) = queue.flow_bucket_items[bucket].front() {
-            // Bucket still has packets. Advance head-finish to
-            // the NEW head packet's finish: head += bytes(new head).
-            // This is the "fresh HOL key" for the next min-scan;
-            // without it, the bucket's selection key would stay
-            // frozen at the just-popped packet's finish and
-            // equal-depth backlogged flows would drain in
-            // `A,A,B,B` bursts (Codex HIGH on the first Phase 3
-            // revision).
-            let next_bytes = cos_item_len(next_head);
-            queue.flow_bucket_head_finish_bytes[bucket] = queue.flow_bucket_head_finish_bytes
-                [bucket]
-                .saturating_add(next_bytes);
-        } else {
-            // Bucket drained — deregister from the active set.
-            // `FlowRrRing::remove` is O(active_count), typically
-            // 2-16 compares; bounded by 1024 worst case.
-            queue.flow_rr_buckets.remove(bucket_u16);
-        }
-        item
-    };
-    // #774: decrement the Local counter BEFORE account_flow_dequeue
-    // so that if account_flow_dequeue panics the counter isn't
-    // stuck high. saturating_sub is a no-op on 0 (never should be
-    // 0 when a Local item is popping, but defense-in-depth).
-    if matches!(item, CoSPendingTxItem::Local(_)) {
-        queue.local_item_count = queue.local_item_count.saturating_sub(1);
-    }
-    let item_len = cos_item_len(&item);
-    let flow_key = cos_item_flow_key(&item);
-    account_cos_queue_flow_dequeue(queue, flow_key, item_len);
-    Some(item)
-}
